Restructure test to reflect changes
