+++ tests/image/test_tiff.py
